Enhance Docker image management in install.sh
Updated the script to check for a local versions file and compare it with the latest from the FTP server. If updates are available, new Docker images are downloaded and the local versions file is updated. If the local file is missing, the latest versions file and corresponding images are downloaded. Removed the previous method of downloading specific images and ensured the latest versions file is saved locally.